// src/types/businessMap.ts
export interface BusinessMapData {
  empresa: string;
  missao: string;
  visao: string;
  valores: string;
  parceirosChave: string;
  atividadesChaves: string;
  recursosChave: string;
  valuePropositions: string;
  relacaoConsumidor: string;
  canaisDistribuicao: string;
  segmentoConsumidores: string;
  estruturaCustos: string;
  fontesReceita: string;
  vantagemCompetitiva: string;
  competenciasEssenciais: string;
  posicionamentoMercado: string;
}